mod: replace warnings with logs
